requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
httpx==0.28.1
anthropic==0.101.0
supabase==2.30.0
APScheduler==3.11.2
python-dotenv==1.2.2
requests==2.33.1
